requirements.txt
flask==2.0.1
flask-cors==3.0.10
openai==1.3.0
python-dotenv==0.19.0
SpeechRecognition==3.8.1
pyaudio==0.2.11
numpy==1.21.2
transformers==4.11.3
torch==1.9.0 